+++ gui/nad_maps_popup_base.ui
